feat: Make integer conversion more correct
An integer contain either a commas or decimal points as the thousand
separator, not both. Thus, replace only one of the two. Furthermore,
depending on the length of the string representation of the number it
may only contain a certain number of thousand separators. Replace max.
the allowed number of thousand separators.